core-api/requirements.txt
# Core Dependencies
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.0.0
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Statistical Analysis (for future projections)
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

